+++ src/GarbageCollector/GarbageCollector.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace Ingenerator\KohanaExtras\GarbageCollector;
+
+
+interface GarbageCollector
+{
+
+    public function execute(): void;
+}