+++ test/mock/Session/MysqlSessionStub.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace test\mock\Ingenerator\KohanaExtras\Session;
+
+
+use Ingenerator\PHPUtils\Session\MysqlSession;
+
+class MysqlSessionStub extends MysqlSession
+{
+    public function __construct()
+    {
+    }
+
+}