+++ test/unit/DependencyFactory/AbstractDependencyFactoryTest.php
@@ -6,11 +6,20 @@
 
 namespace test\unit\Ingenerator\KohanaExtras\DependencyFactory;
 
-use Ingenerator\KohanaExtras\DependencyFactory\KohanaCoreFactory;
 use Ingenerator\KohanaExtras\DependencyFactory\MissingOptionalDependencyException;
 
 abstract class AbstractDependencyFactoryTest extends \PHPUnit\Framework\TestCase
 {
+    /**
+     * A map of service_name => stub class name for shared services that our definitions depend on.
+     *
+     * For example:
+     *   $this->stub_services = ['kohana.psr_log' => NullLogger::class]
+     *
+     * These will be made available to anything that calls ->assertDefinesService
+     */
+    protected array $stub_services = [];
+
     /**
      * @param string $service     key to locate
      * @param array  $definitions the array of definitions provided by this factory
@@ -19,7 +28,8 @@ abstract class AbstractDependencyFactoryTest extends \PHPUnit\Framework\TestCase
      */
     protected function assertDefinesService($service, array $definitions)
     {
-        $list = \Dependency_Definition_List::factory()->from_array($definitions);
+        $definitions = $this->mergeStubServices($definitions);
+        $list        = \Dependency_Definition_List::factory()->from_array($definitions);
         try {
             $list->get($service);
         } catch (\Exception $e) {
@@ -60,4 +70,20 @@ abstract class AbstractDependencyFactoryTest extends \PHPUnit\Framework\TestCase
             }
         }
     }
+
+    /**
+     * @param array $definitions
+     *
+     * @return array
+     */
+    protected function mergeStubServices(array $definitions): array
+    {
+        // Merges in $this->stub_services
+        $stubs = [];
+        foreach ($this->stub_services as $s => $class) {
+            \Arr::set_path($stubs, $s, ['_settings' => ['class' => $class]]);
+        }
+
+        return \Arr::merge($stubs, $definitions);
+}
 }